% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esp.R
\name{esp}
\alias{esp}
\title{enhanced stratified power}
\usage{
esp(
  formula,
  data,
  listw = NULL,
  discvar = "all",
  discnum = 3:8,
  model = "ols",
  Durbin = FALSE,
  overlay = "and",
  alpha = 0.75,
  bw = "AIC",
  adaptive = TRUE,
  kernel = "gaussian",
  cores = 1,
  ...
)
}
\arguments{
\item{formula}{A formula for enhanced stratified power model.}

\item{data}{An \code{sf} object of observation data. Please note that the column names of the independent
variables should not be \code{all} or \code{none}.}

\item{listw}{(optional) A \code{listw}. See \code{spdep::mat2listw()} and \code{spdep::nb2listw()} for details.}

\item{discvar}{(optional) Name of continuous variable columns that need to be discretized. Noted that
when \code{formula} has \code{discvar}, \code{data} must have these columns. Default is \code{all}, which means all independent
variables are used as \code{discvar}. When \code{discvar} is set to \code{none}, all independent variables do not need to
be discretized.}

\item{discnum}{(optional) Number of discretization. Default all will use \code{3:8}.}

\item{model}{(optional) The type of linear model used, default is \code{ols}. The \code{model} value must be any of
\code{ols}, \code{gwr}, \code{lag} or \code{error}.}

\item{Durbin}{(optional) Whether to consider spatial Durbin terms, default is \code{false}.}

\item{overlay}{(optional) Spatial overlay method. One of \code{and}, \code{or}, \code{intersection}. Default is \code{and}.}

\item{alpha}{(optional) Controlling the strength of spatial soft constraints, the larger the \code{alpha},
the stronger the spatial soft constraint. Default is \code{0.75}.}

\item{bw}{(optional) The bandwidth used in selecting models. The optimal bandwidth can be
selected using one of two methods: \code{AIC}, and \code{CV}. Default will use \code{AIC}.}

\item{adaptive}{(optional) Whether the bandwidth value is adaptive or not. Default is \code{TRUE}.}

\item{kernel}{(optional) Kernel function. Default is \code{gaussian}.}

\item{cores}{(optional) Positive integer (default is 1). When cores are greater than 1, use
multi-core parallel computing.}

\item{...}{(optional) Other arguments passed to \code{ClustGeo::hclustgeo()}.}
}
\value{
A list.
\describe{
\item{\code{factor}}{factor detection result}
\item{\code{interaction}}{Interactive detection results}
\item{\code{pred}}{predicted values of response variables required for the calculation of q-values}
\item{\code{disc}}{spatial discretization results}
\item{\code{y}}{values corresponding to the response variable}
\item{\code{xvar}}{all the explanatory variables}
\item{\code{allvar}}{all explanatory variables and their pairwise interaction variables.}
}
}
\description{
Equivalent geographical detector q-statistic under a spatial linear regression framework.
}
\examples{
\dontrun{
depression = system.file('extdata/Depression.csv',package = 'gdverse') |>
  readr::read_csv() |>
  sf::st_as_sf(coords = c('X','Y'), crs = 4326)
g = esp(Depression_prevelence ~ ., data = depression, cores = 6)
g$factor
}
}
