% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_betas.R
\name{gwr_betas}
\alias{gwr_betas}
\title{Estimate \code{gwr} model coefficients}
\usage{
gwr_betas(
  formula,
  data,
  bw = "AIC",
  adaptive = TRUE,
  kernel = "gaussian",
  intercept = FALSE
)
}
\arguments{
\item{formula}{A formula of \code{gwr} model.}

\item{data}{An sf object of observation data}

\item{bw}{(optional) The bandwidth used in selecting models. The optimal bandwidth can be
selected using one of two methods: \code{AIC}, and \code{CV}. Default will use \code{AIC}.}

\item{adaptive}{(optional) Whether the bandwidth value is adaptive or not. Default is \code{TRUE}.}

\item{kernel}{(optional) Kernel function. Default is \code{gaussian}.}

\item{intercept}{(optional) Whether to include the intercept term in the coefficient \verb{tibble. Default is }FALSE`.}
}
\value{
A \code{tibble}
}
\description{
Estimate \code{gwr} model coefficients
}
\examples{
\donttest{
depression = system.file('extdata/Depression.csv',package = 'gdverse') |>
  readr::read_csv() |>
  sf::st_as_sf(coords = c('X','Y'), crs = 4326)
gwr_betas(Depression_prevelence ~ ., data = depression)
gwr_betas(Depression_prevelence ~ ., data = depression, intercept = TRUE)
}
}
