% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesp.R
\name{sesp}
\alias{sesp}
\title{Spatially Explicit Stratified Power (SESP) Model}
\usage{
sesp(
  formula,
  data,
  listw = NULL,
  discvar = "all",
  discnum = 3:8,
  model = "ols",
  durbin = FALSE,
  overlay = "and",
  alpha = 0.5,
  intercept = FALSE,
  bw = "AIC",
  adaptive = TRUE,
  kernel = "gaussian",
  increase_rate = 0.05,
  cores = 1,
  ...
)
}
\arguments{
\item{formula}{A formula for enhanced stratified power model.}

\item{data}{An \code{sf} object of observation data. Please note that the column names of the independent
variables should not be \code{all} or \code{none}.}

\item{listw}{(optional) A \code{listw} object. See \code{spdep::mat2listw()} and \code{spdep::nb2listw()} for details.}

\item{discvar}{(optional) Name of continuous variable columns that need to be discretized. Noted that
when \code{formula} has \code{discvar}, \code{data} must have these columns. Default is \code{all}, which means all independent
variables are used as \code{discvar}. When \code{discvar} is set to \code{none}, all independent variables do not need to
be discretized.}

\item{discnum}{(optional) Number of discretization. Default all will use \code{3:8}.}

\item{model}{(optional) The type of linear model used, default is \code{ols}. The \code{model} value must be any of
\code{ols}, \code{gwr}, \code{lag} or \code{error}.}

\item{durbin}{(optional) Whether to consider spatial durbin terms, default is \code{false}.}

\item{overlay}{(optional) Spatial overlay method. One of \code{and}, \code{or}, \code{intersection}. Default is \code{and}.}

\item{alpha}{(optional) Controlling the strength of spatial soft constraints, the larger the \code{alpha},
the stronger the spatial soft constraint. Default is \code{0.5}.}

\item{intercept}{(optional) Whether to include the intercept term in the gwr coefficient \code{tibble}.
Default is \code{FALSE}.}

\item{bw}{(optional) The bandwidth used in selecting models. The optimal bandwidth can be
selected using one of two methods: \code{AIC}, and \code{CV}. Default is \code{AIC}.}

\item{adaptive}{(optional) Whether the bandwidth value is adaptive or not. Default is \code{TRUE}.}

\item{kernel}{(optional) Kernel function. Default is \code{gaussian}.}

\item{increase_rate}{(optional) The critical increase rate of the number of discretization.
Default is \verb{5\%}.}

\item{cores}{(optional) Positive integer (default is 1). When cores are greater than 1, use
multi-core parallel computing.}

\item{...}{(optional) Other arguments passed to \code{sdsfun::hclustgeo_disc()}.}
}
\value{
A list.
\describe{
\item{\code{factor}}{global factor detection result}
\item{\code{interaction}}{global interactive detection results}
\item{\code{optdisc}}{independent variable optimal spatial discretization}
\item{\code{allfactor}}{factor detection results corresponding to different numbers of discreteization}
\item{\code{model}}{regression model used to estimate equivalent q values}
}
}
\description{
Spatially Explicit Stratified Power (SESP) Model
}
\note{
Note that when the number of continuous independent variables is small (three or fewer), the built-in spatial explicit
discretization in \code{sesp} may overestimate the variable q value (when there are fewer input independent variables, GWR
tends to overestimate the contribution of each variable). In such cases, it is recommended to discretize these variables
beforehand and then input them into \code{sesp} for computation.
}
\examples{
NTDs = sf::st_as_sf(gdverse::NTDs, coords = c('X','Y'))
g = sesp(incidence ~ ., data = NTDs, discvar = 'none',
         model = 'ols', overlay = 'intersection', cores = 1)
g

}
