% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esp.R
\name{esp}
\alias{esp}
\title{enhanced stratified power}
\usage{
esp(
  formula,
  data,
  listw = NULL,
  yzone = NULL,
  discvar = "all",
  discnum = 3:8,
  model = "ols",
  Durbin = FALSE,
  overlay = "and",
  alpha = 0.75,
  bw = "AIC",
  adaptive = TRUE,
  kernel = "gaussian",
  increase_rate = 0.05,
  cores = 1,
  ...
)
}
\arguments{
\item{formula}{A formula for enhanced stratified power model.}

\item{data}{An \code{sf} object of observation data. Please note that the column names of the independent
variables should not be \code{all} or \code{none}.}

\item{listw}{(optional) A list of \code{listw}. See \code{spdep::mat2listw()} and \code{spdep::nb2listw()} for details.}

\item{yzone}{(optional) Spatial zones of the response variable. Default the response variable is divided into
\code{3} categories using \code{quantile} discretization.}

\item{discvar}{(optional) Name of continuous variable columns that need to be discretized. Noted that
when \code{formula} has \code{discvar}, \code{data} must have these columns. Default is \code{all}, which means all independent
variables are used as \code{discvar}. When \code{discvar} is set to \code{none}, all independent variables do not need to
be discretized.}

\item{discnum}{(optional) Number of discretization. Default all will use \code{3:8}.}

\item{model}{(optional) The type of linear model used, default is \code{ols}. The \code{model} value must be any of
\code{ols}, \code{gwr}, \code{lag} or \code{error}.}

\item{Durbin}{(optional) Whether to consider spatial Durbin terms, default is \code{false}.}

\item{overlay}{(optional) Spatial overlay method. One of \code{and}, \code{or}, \code{intersection}. Default is \code{and}.}

\item{alpha}{(optional) Controlling the strength of spatial soft constraints, the larger the \code{alpha},
the stronger the spatial soft constraint. Default is \code{0.75}.}

\item{bw}{(optional) The bandwidth used in selecting models. The optimal bandwidth can be
selected using one of two methods: \code{AIC}, and \code{CV}. Default will use \code{AIC}.}

\item{adaptive}{(optional) Whether the bandwidth value is adaptive or not. Default is \code{TRUE}.}

\item{kernel}{(optional) Kernel function. Default is \code{gaussian}.}

\item{increase_rate}{(optional) The critical increase rate of the number of discretization.
Default is \verb{5\%}.}

\item{cores}{(optional) Positive integer (default is 1). When cores are greater than 1, use
multi-core parallel computing.}

\item{...}{(optional) Other arguments passed to \code{ClustGeo::hclustgeo()}.}
}
\value{
A list with \code{espm} class.
\describe{
\item{\code{factor}}{global factor detection result}
\item{\code{interaction}}{global interactive detection results}
\item{\code{optdisc}}{independent variable optimal spatial discretization}
\item{\code{localq}}{q values of explanatory variables under different response zones}
\item{\code{zone}}{zones of the response variable}
\item{\code{allfactor}}{factor detection results corresponding to different numbers of discreteization}
\item{\code{model}}{regression model used to estimate equivalent q values}
}
}
\description{
Equivalent geographical detector q-statistic under a spatial linear regression framework.
}
\examples{
NTDs = sf::st_as_sf(gdverse::NTDs, coords = c('X','Y'))
g = esp(incidence ~ ., data = NTDs, discvar = 'none',
        model = 'ols', overlay = 'intersection', cores = 1)
g

}
